requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
Werkzeug==1.0.1
wincertstore==0.2
